Update cd_frame_generator.cpp
Given that the events are accumulated into frames where the accumulation windows are overlapping in nature, the "fps" and "display_accumulation_time_us" ideally should be decoupled from each other.  The variable "notify_slice_us_", conditional on the value of "display_accumulation_time_us", seems to affect the output framerate in proportion to the event accumulation time. 
For example, if the fps is set to 100Hz, and the accumulation time is set to 50000us, next_notify_us_ is set to 50000/3 = 1666.67, making the resultant output framerate 60Hz even when CPU cores are available.
The proposed change to set notify_slice_us_ to a constant 100us attempts to decouple output frame rate from event accumulation times.